+++ packages/react/src/exampleTypes/Number.js
@@ -24,7 +24,7 @@ let NumberComponent = injectTreeNode(
             onChange={e =>
               tree.mutate(node.path, {
                 min: _.isNumber(significantDigits)
-                  ? round(e.target.value, significantDigits)
+                  ? _.toString(round(e.target.value, significantDigits))
                   : e.target.value,
               })
             }
@@ -35,7 +35,7 @@ let NumberComponent = injectTreeNode(
             onChange={e =>
               tree.mutate(node.path, {
                 max: _.isNumber(significantDigits)
-                  ? round(e.target.value, significantDigits)
+                  ? _.toString(round(e.target.value, significantDigits))
                   : e.target.value,
               })
             }