+++ packages/react/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.54.9
+* TermStatsTable: Add as filter would now add filter if the existing one is not in mode 'include'
+* Number: Better handling zero edge case when significantDigits is true
+* Facet: Better handling of the labels when there is no white space
+
 # 1.54.8
 * TagsInput: Better validation and input logic for new tags